move angle, norm, real, imag to individual modules
