+++ src/phasor.rs
@@ -1,5 +1,4 @@
 use crate::trig::*;
-use core::f64::consts::PI;
 
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
@@ -8,10 +7,14 @@ use wasm_bindgen::prelude::*;
 use proptest_derive::Arbitrary;
 
 mod add;
+mod angle;
 mod approx;
 mod display;
 mod div;
+mod imag;
 mod mul;
+mod norm;
+mod real;
 mod sub;
 
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
@@ -45,68 +48,6 @@ impl Phasor {
             } else {
                 im / re
             },
-        }
-    }
-
-    #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
-    pub fn real(&self) -> f64 {
-        self.mag * cosatan(self.tan)
-    }
-
-    #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
-    pub fn imag(&self) -> f64 {
-        self.mag * sinatan(self.tan)
-    }
-
-    #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
-    pub fn norm(&self) -> f64 {
-        self.mag.abs()
-    }
-
-    #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
-    pub fn angle(&self) -> f64 {
-        if self.mag.is_sign_positive() {
-            self.tan.atan()
-        } else {
-            self.tan.atan() - PI.copysign(self.tan)
-        }
-    }
-}
-
-#[cfg(all(test, not(target_arch = "wasm32")))]
-mod tests {
-    use super::*;
-    use crate::assert_close_to;
-    use alloc::format;
-    use proptest::prelude::*;
-
-    proptest! {
-        #[test]
-        fn real(re: f64, im: f64) {
-            let p = Phasor::rect(re as f64, im as f64);
-            prop_assume!(p.real().is_normal());
-            assert_close_to!(p.real(), re as f64);
-        }
-
-        #[test]
-        fn imag(re: f64, im: f64) {
-            let p = Phasor::rect(re as f64, im as f64);
-            prop_assume!(p.real().is_normal());
-            assert_close_to!(p.imag(), im as f64);
-        }
-
-        #[test]
-        fn norm(mag: f64, ang: f64) {
-            let p = Phasor::polar(mag, ang);
-            assert_close_to!(p.norm(), mag.abs());
-        }
-
-        #[test]
-        fn angle(mag: f64, ang: f64) {
-            let p = Phasor::polar(mag, ang);
-
-            assert_close_to!(p.angle().cos(), ang.cos() * mag.signum());
-            assert_close_to!(p.angle().sin(), ang.sin() * mag.signum());
         }
     }
 }
